Question APIs and Repository Updated

+++ src/main/java/com/exam/question/repository/QuestionRepository.java
@@ -2,6 +2,9 @@ package com.exam.question.repository;
 
 import com.exam.common.repository.BaseRepository;
 import com.exam.question.model.Question;
+import com.exam.question.model.QuestionChoice;
+import com.exam.question.model.QuestionDataItem;
+import com.exam.question.repository.rowmappers.QuestionDataItemRowMapper;
 import com.exam.question.repository.rowmappers.QuestionRowMapper;
 import org.springframework.dao.DataAccessException;
 import org.springframework.dao.EmptyResultDataAccessException;
@@ -9,27 +12,67 @@ import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
 import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.stereotype.Repository;
+import org.springframework.transaction.annotation.Transactional;
 
+import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 @Repository
 public class QuestionRepository extends BaseRepository {
 
-    public static final StringBuilder FIND_ALL_QUESTION_QUERY = new StringBuilder("SELECT * FROM questions");
+    public static final StringBuilder FIND_ALL_QUESTION_QUERY = new StringBuilder("SELECT * FROM questions JOIN ");
     public static final StringBuilder DELETE_ALL_QUESTION_QUERY = new StringBuilder("DELETE FROM questions");
 
     //CREATE QUERIES
 
+    @Transactional
     public int addQuestion(Question question, int loggedInUserId){
-        final StringBuilder sql = new StringBuilder("INSERT INTO questions(title, description, enabled, topic_id, created_by, modified_by)");
-        sql.append(" VALUES (:title,:description,:enabled,:topicId, :createdBy, :modifiedBy");
+        final StringBuilder sql = new StringBuilder("INSERT INTO questions(title, description, enabled, proficiency_id, topic_id, created_by, modified_by)");
+        sql.append(" VALUES (:title, :description, :enabled, :proficiencyId, :topicId, :createdBy, :modifiedBy");
         sql.append(") RETURNING id");
 
         MapSqlParameterSource param = new MapSqlParameterSource();
         param.addValue("title", question.getTitle());
         param.addValue("description", question.getDescription());
         param.addValue("enabled", question.isEnabled());
+        param.addValue("proficiencyId", question.getProficiencyId());
         param.addValue("topicId", question.getTopicId());
+        param.addValue("createdBy", loggedInUserId);
+        param.addValue("modifiedBy", loggedInUserId);
+
+        try{
+            GeneratedKeyHolder generatedKeyHolder = new GeneratedKeyHolder();
+            npJdbcTemplate.update(sql.toString(), param, generatedKeyHolder);
+            int questionId = generatedKeyHolder.getKey().intValue();
+
+            if(questionId > 0 && question.getQuestionChoices() != null && question.getQuestionChoices().size() > 0){
+                question.getQuestionChoices().forEach(questionChoice -> {
+                    questionChoice.setQuestionId(questionId);
+                    int choiceId = addQuestionChoice(questionChoice, loggedInUserId);
+                    questionChoice.setId(choiceId);
+                });
+            }
+
+            return questionId;
+
+        }catch (EmptyResultDataAccessException e){
+            return -1;
+        }
+    }
+
+    private int addQuestionChoice(QuestionChoice questionChoice, int loggedInUserId) {
+        final StringBuilder sql = new StringBuilder("INSERT INTO exam.question_choices ");
+        sql.append("(description, enabled, correct, ques_id, created_by, modified_by)");
+        sql.append("VALUES(:description, :enabled, :correct, :ques_id, :createdBy, :modifiedBy)");
+        sql.append(" RETURNING id");
+
+        MapSqlParameterSource param = new MapSqlParameterSource();
+        param.addValue("description", questionChoice.getDescription());
+        param.addValue("enabled", questionChoice.isEnabled());
+        param.addValue("correct", questionChoice.isCorrect());
+        param.addValue("ques_id", questionChoice.getQuestionId());
         param.addValue("createdBy", loggedInUserId);
         param.addValue("modifiedBy", loggedInUserId);
 
@@ -46,20 +89,38 @@ public class QuestionRepository extends BaseRepository {
     //SELECT QUERIES
 
     public Question findById(int id){
-        final String sql = "SELECT * FROM questions WHERE id=:id";
+        final StringBuilder sql = new StringBuilder("SELECT q.* , ");
+        sql.append("qc.id as qc_id, ");
+        sql.append("qc.description as qc_desc, ");
+        sql.append("qc.enabled as qc_enabled, ");
+        sql.append("qc.correct as qc_correct ");
+        sql.append("FROM questions q ");
+        sql.append("JOIN question_choices qc on qc.ques_id = q.id ");
+        sql.append("WHERE q.id=:id ");
         final SqlParameterSource param = new MapSqlParameterSource("id",id);
         try{
-            return (Question) npJdbcTemplate.queryForObject(sql, param, new QuestionRowMapper());
+            List<QuestionDataItem> dataItems = (List<QuestionDataItem>) npJdbcTemplate.query(sql.toString(), param, new QuestionDataItemRowMapper());
+            List<Question> questions = new ArrayList<>(convertQuesDataItemsToQuestionList(dataItems).values());
+            return questions.get(0);
         }catch (EmptyResultDataAccessException e){
             return null;
         }
     }
 
     public Question findByTitle(String title){
-        final String sql = "SELECT * FROM questions WHERE title=:title";
+        final StringBuilder sql = new StringBuilder("SELECT q.* , ");
+        sql.append("qc.id as qc_id, ");
+        sql.append("qc.description as qc_desc, ");
+        sql.append("qc.enabled as qc_enabled, ");
+        sql.append("qc.correct as qc_correct ");
+        sql.append("FROM questions q ");
+        sql.append("JOIN question_choices qc on qc.ques_id = q.id ");
+        sql.append("WHERE title=:title");
         final SqlParameterSource param = new MapSqlParameterSource("title",title);
         try{
-            return (Question) npJdbcTemplate.queryForObject(sql, param, new QuestionRowMapper());
+            List<QuestionDataItem> dataItems = (List<QuestionDataItem>) npJdbcTemplate.query(sql.toString(), param, new QuestionDataItemRowMapper());
+            List<Question> questions = new ArrayList<>(convertQuesDataItemsToQuestionList(dataItems).values());
+            return questions.get(0);
         }catch (EmptyResultDataAccessException e){
             return null;
         }
@@ -67,10 +128,55 @@ public class QuestionRepository extends BaseRepository {
 
     public List<Question> findAll(){
         try{
-            return npJdbcTemplate.query(FIND_ALL_QUESTION_QUERY.toString(), new QuestionRowMapper());
+            final StringBuilder sql = new StringBuilder("SELECT q.* , ");
+            sql.append("qc.id as qc_id, ");
+            sql.append("qc.description as qc_desc, ");
+            sql.append("qc.enabled as qc_enabled, ");
+            sql.append("qc.correct as qc_correct ");
+            sql.append("FROM questions q ");
+            sql.append("JOIN question_choices qc on qc.ques_id = q.id ");
+            List<QuestionDataItem> dataItems = (List<QuestionDataItem>) npJdbcTemplate.query(sql.toString(), new QuestionDataItemRowMapper());
+            return new ArrayList<>(convertQuesDataItemsToQuestionList(dataItems).values());
         }catch (DataAccessException e){
             return null;
         }
+    }
+
+    private Map<Integer, Question> convertQuesDataItemsToQuestionList(List<QuestionDataItem> dataItemList) {
+        Map<Integer, Question> questionMap = new HashMap<>();
+        for(QuestionDataItem item : dataItemList) {
+            Integer quesId = item.getId();
+            Question question = questionMap.get(quesId);
+            if(question == null){
+                question = new Question();
+                question.setId(item.getId());
+                question.setTitle(item.getTitle());
+                question.setDescription(item.getDescription());
+                question.setEnabled(item.isEnabled());
+                question.setProficiencyId(item.getProficiencyId());
+                question.setTopicId(item.getTopicId());
+
+                List<QuestionChoice> choices = new ArrayList<>();
+                if(item.getQuestionChoice() != null){
+                    item.getQuestionChoice().setQuestionId(question.getId());
+                }
+                choices.add(item.getQuestionChoice());
+                question.setQuestionChoices(choices);
+            }else {
+                List<QuestionChoice> choices = question.getQuestionChoices();
+                if(choices == null) {
+                    choices = new ArrayList<>();
+                }
+                if(item.getQuestionChoice() != null){
+                    item.getQuestionChoice().setQuestionId(question.getId());
+                }
+                choices.add(item.getQuestionChoice());
+                question.setQuestionChoices(choices);
+            }
+
+            questionMap.put(question.getId(), question);
+        }
+        return questionMap;
     }
 
     public int findTotalCount(){
@@ -105,28 +211,73 @@ public class QuestionRepository extends BaseRepository {
 
     //DELETE QUERIES
 
+    @Transactional
     public boolean delete(int id) {
-        final String sql = "DELETE FROM questions WHERE id=:id";
-        final SqlParameterSource param = new MapSqlParameterSource("id",id);
+        boolean choiceDeleteStatus = deleteQuestionChoiceByQuestionId(id);
+        if(choiceDeleteStatus){
+            final String sql = "DELETE FROM questions WHERE id=:id";
+            final SqlParameterSource param = new MapSqlParameterSource("id",id);
+            try{
+                return npJdbcTemplate.update(sql, param) > 0;
+            }catch (DataAccessException e){
+                return false;
+            }
+        }
+        return false;
+    }
+
+    private boolean deleteQuestionChoiceByQuestionId(int quesId) {
+        final String sql = "DELETE FROM question_choices qc WHERE qc.ques_id=:quesId";
+        final SqlParameterSource param = new MapSqlParameterSource("quesId", quesId);
         try{
             return npJdbcTemplate.update(sql, param) > 0;
-        }catch (DataAccessException e){
+        }catch(DataAccessException e){
             return false;
         }
     }
 
+    @Transactional
     public boolean deleteByIds(List<Integer> ids){
-        final String sql = "DELETE FROM questions WHERE id in (:ids)";
-        final SqlParameterSource param = new MapSqlParameterSource().addValue("ids", ids);
+        boolean choicesDeleteStatus = deleteQuestionChoicesByQuestionIds(ids);
+        if(choicesDeleteStatus) {
+            final String sql = "DELETE FROM questions WHERE id in (:ids)";
+            final SqlParameterSource param = new MapSqlParameterSource().addValue("ids", ids);
+            try {
+                return npJdbcTemplate.update(sql, param) > 0;
+            } catch (DataAccessException e) {
+                return false;
+            }
+        }
+        return false;
+    }
+
+    private boolean deleteQuestionChoicesByQuestionIds(List<Integer> quesIds) {
+        final String sql = "DELETE FROM question_choices qc WHERE qc.ques_id in (:quesIds)";
+        final SqlParameterSource param = new MapSqlParameterSource("quesIds", quesIds);
         try{
             return npJdbcTemplate.update(sql, param) > 0;
-        }catch (DataAccessException e){
+        }catch(DataAccessException e){
             return false;
         }
     }
 
+    @Transactional
     public boolean deleteAll(){
-        final String sql = DELETE_ALL_QUESTION_QUERY.toString();
+        boolean choicesDeleteStatus = deleteAllQuestionChoicesForAllQuestions();
+        if(choicesDeleteStatus) {
+            final String sql = DELETE_ALL_QUESTION_QUERY.toString();
+            final SqlParameterSource param = new MapSqlParameterSource();
+            try {
+                return npJdbcTemplate.update(sql, param) > 0;
+            } catch (DataAccessException e) {
+                return false;
+            }
+        }
+        return false;
+    }
+
+    private boolean deleteAllQuestionChoicesForAllQuestions(){
+        final String sql = "DELETE FROM question_choices";
         final SqlParameterSource param = new MapSqlParameterSource();
         try{
             return npJdbcTemplate.update(sql, param) > 0;
@@ -137,26 +288,90 @@ public class QuestionRepository extends BaseRepository {
 
     //UPDATE QUERIES
 
+    @Transactional
     public boolean updateQuestion(int id, Question question, int loggedInUserId){
-        final StringBuilder sql = new StringBuilder("UPDATE questions SET ");
-        sql.append("title=:title,");
+        Question quesOldData = findById(id);
+        List<QuestionChoice> oldChoices = quesOldData.getQuestionChoices();
+        List<QuestionChoice> newChoices = question.getQuestionChoices();
+
+        boolean choicesModified = true;
+        if(oldChoices != null){
+            choicesModified = !oldChoices.equals(newChoices);
+        }
+
+        boolean choicesUpdateStatus = true;
+        if(choicesModified && newChoices != null){
+            boolean oldChoicesDeleted = deleteQuestionChoiceByQuestionId(id);
+            if(oldChoicesDeleted){
+                for (QuestionChoice newChoice : newChoices) {
+                    newChoice.setQuestionId(id);
+                    int choiceId = addQuestionChoice(newChoice, loggedInUserId);
+                    newChoice.setId(choiceId);
+                    choicesUpdateStatus = choiceId > 0;
+                }
+            }
+        }
+
+        if(choicesUpdateStatus) {
+            final StringBuilder sql = new StringBuilder("UPDATE questions SET ");
+            sql.append("title=:title,");
+            sql.append("description=:description,");
+            sql.append("enabled=:enabled,");
+            sql.append("proficiency_id=:proficiencyId, ");
+            sql.append("topic_id=:topicId, ");
+            sql.append("modified_by=:loggedInUserId ");
+            sql.append("WHERE id=:id");
+
+            MapSqlParameterSource param = new MapSqlParameterSource("id", id)
+                    .addValue("title", question.getTitle())
+                    .addValue("description", question.getDescription())
+                    .addValue("enabled", question.isEnabled())
+                    .addValue("proficiencyId", question.getProficiencyId())
+                    .addValue("topicId", question.getTopicId())
+                    .addValue("id", id)
+                    .addValue("loggedInUserId", loggedInUserId);
+            try {
+                return npJdbcTemplate.update(sql.toString(), param) > 0;
+            } catch (DataAccessException e) {
+                return false;
+            }
+        }
+        return false;
+    }
+
+    private boolean updateQuestionChoices(int choiceId, QuestionChoice choice, int loggedInUserId) {
+        final StringBuilder sql = new StringBuilder("UPDATE question_choices SET ");
         sql.append("description=:description,");
         sql.append("enabled=:enabled,");
-        sql.append("topic_id=:topicId ");
+        sql.append("correct=:correct,");
         sql.append("modified_by=:loggedInUserId ");
-        sql.append("WHERE id=:id");
+        sql.append("WHERE id=:choiceId");
 
-        MapSqlParameterSource param = new MapSqlParameterSource("id",id)
-                .addValue("title", question.getTitle())
-                .addValue("description", question.getDescription())
-                .addValue("enabled", question.isEnabled())
-                .addValue("topicId", question.getTopicId())
-                .addValue("id", id)
+        MapSqlParameterSource param = new MapSqlParameterSource()
+                .addValue("description", choice.getDescription())
+                .addValue("enabled", choice.isEnabled())
+                .addValue("correct", choice.isCorrect())
+                .addValue("choiceId", choiceId)
                 .addValue("loggedInUserId", loggedInUserId);
-        try{
+        try {
             return npJdbcTemplate.update(sql.toString(), param) > 0;
-        }catch (DataAccessException e){
+        } catch (DataAccessException e) {
             return false;
         }
+    }
+
+    @Transactional
+    public boolean updateQuestionChoicesByQuestionId(int quesId, List<QuestionChoice> questionChoices, int loggedInUserId) {
+        boolean choicesUpdateStatus = true;
+        boolean oldChoicesDeleted = deleteQuestionChoiceByQuestionId(quesId);
+        if(oldChoicesDeleted){
+            for (QuestionChoice newChoice : questionChoices) {
+                newChoice.setQuestionId(quesId);
+                int choiceId = addQuestionChoice(newChoice, loggedInUserId);
+                newChoice.setId(choiceId);
+                choicesUpdateStatus = choiceId > 0;
+            }
+        }
+        return choicesUpdateStatus;
     }
 }
